+++ app/models/audit.py
@@ -2,14 +2,14 @@ from __future__ import annotations
 from datetime import datetime
 from typing import Optional
 from sqlalchemy.orm import Mapped, mapped_column
-from sqlalchemy import BigInteger, String, DateTime, Text, ForeignKey, CheckConstraint
+from sqlalchemy import Integer, String, DateTime, Text, ForeignKey, CheckConstraint
 from . import Base
 
 
 class AuditLog(Base):
     __tablename__ = "audit_logs"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     actor_type: Mapped[str] = mapped_column(String(20), nullable=False)
     actor_admin_id: Mapped[Optional[int]] = mapped_column(
         ForeignKey("admins.id", ondelete="SET NULL"), nullable=True
@@ -19,7 +19,7 @@ class AuditLog(Base):
     )
     action: Mapped[str] = mapped_column(String(100), nullable=False)
     subject_table: Mapped[str] = mapped_column(String(50), nullable=False)
-    subject_id: Mapped[Optional[int]] = mapped_column(BigInteger, nullable=True)
+    subject_id: Mapped[Optional[int]] = mapped_column(Integer, nullable=True)
     details_json: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
     occurred_at: Mapped[datetime] = mapped_column(DateTime, nullable=False)
