+++ app/models/chain.py
@@ -3,7 +3,7 @@ from datetime import datetime
 from typing import TYPE_CHECKING, Optional
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from sqlalchemy import (
-    BigInteger,
+    Integer,
     String,
     DateTime,
     Text,
@@ -21,7 +21,7 @@ if TYPE_CHECKING:
 class BlockchainTransaction(Base):
     __tablename__ = "blockchain_transactions"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     user_id: Mapped[Optional[int]] = mapped_column(
         ForeignKey("users.id", ondelete="SET NULL"), nullable=True
     )
